+++ plugin/src/main/java/com/stratio/cassandra/lucene/search/SearchBuilder.java
@@ -16,6 +16,7 @@
 package com.stratio.cassandra.lucene.search;
 
 import com.stratio.cassandra.lucene.IndexException;
+import com.stratio.cassandra.lucene.IndexPagingState;
 import com.stratio.cassandra.lucene.search.condition.Condition;
 import com.stratio.cassandra.lucene.search.condition.builder.ConditionBuilder;
 import com.stratio.cassandra.lucene.search.sort.Sort;
@@ -24,13 +25,9 @@ import com.stratio.cassandra.lucene.search.sort.builder.SortFieldBuilder;
 import com.stratio.cassandra.lucene.util.Builder;
 import com.stratio.cassandra.lucene.util.ByteBufferUtils;
 import com.stratio.cassandra.lucene.util.JsonSerializer;
-import org.apache.cassandra.db.Clustering;
-import org.apache.cassandra.db.DecoratedKey;
-import org.apache.cassandra.db.marshal.CompositeType;
 import org.codehaus.jackson.annotate.JsonProperty;
 
 import java.io.IOException;
-import java.nio.ByteBuffer;
 
 /**
  * Class for building a new {@link Search}.
@@ -55,11 +52,8 @@ public class SearchBuilder implements Builder<Search> {
     @JsonProperty("refresh")
     private boolean refresh;
 
-    @JsonProperty("after_key")
-    private String afterKey;
-
-    @JsonProperty("after_clustering")
-    private String afterClustering;
+    @JsonProperty("paging")
+    private String paging;
 
     /** Default constructor. */
     SearchBuilder() {
@@ -125,63 +119,12 @@ public class SearchBuilder implements Builder<Search> {
     /**
      * Sets the specified starting partition key.
      *
-     * @param hex a partition key formatted as a hex string
-     * @returnthis builder with the specified partition key
+     * @param pagingState a paging state
+     * @return this builder with the specified partition key
      */
-    public SearchBuilder afterKey(String hex) {
-        this.afterKey = hex;
+    public SearchBuilder paging(IndexPagingState pagingState) {
+        this.paging = ByteBufferUtils.toHex(pagingState.toByteBuffer());
         return this;
-    }
-
-    /**
-     * Sets the specified starting partition key.
-     *
-     * @param key a partition key
-     * @returnthis builder with the specified partition key
-     */
-    public SearchBuilder afterKey(ByteBuffer key) {
-        return afterKey(ByteBufferUtils.toHex(key));
-    }
-
-    /**
-     * Sets the specified starting partition key.
-     *
-     * @param key a partition key
-     * @returnthis builder with the specified partition key
-     */
-    public SearchBuilder afterKey(DecoratedKey key) {
-        return afterKey(key.getKey());
-    }
-
-    /**
-     * Sets the specified starting clustering key.
-     *
-     * @param hex a clustering key formatted as a hex string
-     * @return this builder with the specified clustering key
-     */
-    public SearchBuilder afterClustering(String hex) {
-        this.afterClustering = hex;
-        return this;
-    }
-
-    /**
-     * Sets the specified starting clustering key.
-     *
-     * @param clustering a clustering key
-     * @return this builder with the specified clustering key
-     */
-    public SearchBuilder afterClustering(ByteBuffer clustering) {
-        return afterClustering(ByteBufferUtils.toHex(clustering));
-    }
-
-    /**
-     * Sets the specified starting clustering key.
-     *
-     * @param clustering a clustering key
-     * @return this builder with the specified clustering key
-     */
-    public SearchBuilder afterClustering(Clustering clustering) {
-        return afterClustering(CompositeType.build(clustering.getRawValues()));
     }
 
     /**
@@ -193,9 +136,9 @@ public class SearchBuilder implements Builder<Search> {
         return new Search(queryBuilder == null ? null : queryBuilder.build(),
                           filterBuilder == null ? null : filterBuilder.build(),
                           sortBuilder == null ? null : sortBuilder.build(),
-                          refresh,
-                          ByteBufferUtils.byteBuffer(afterKey),
-                          ByteBufferUtils.byteBuffer(afterClustering));
+                          paging == null ? null : IndexPagingState.build(ByteBufferUtils.byteBuffer(paging)),
+                          refresh
+        );
     }
 
     /**