+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/util/CassandraConfig.java
@@ -37,7 +37,7 @@ class CassandraConfig {
     static final String INDEX = getString("index", "test_table_idx");
     static final String COLUMN = getString("column", null);
     static final boolean USE_NEW_QUERY_SYNTAX = getBool("use_new_query_syntax", true);
-    static final int LIMIT = getInt("limit", 10000); // Top-k
+    static final int LIMIT = getInt("limit", 10000);
 
     static {
         assert COLUMN != null || USE_NEW_QUERY_SYNTAX;